@@ -13,7 +13,9 @@ use function is_null;
 use function strtolower;
 use function str_ireplace;
 
-abstract class Element implements JsonSerializable, ArrayAccess   {
+abstract class Element implements JsonSerializable, ArrayAccess, Rendererable   {
+
+    use ArrayTrait;
 
     protected $ast = null;
     /**
@@ -32,7 +34,7 @@ abstract class Element implements JsonSerializable, ArrayAccess   {
         if (is_null($ast)) {
 
             $ast = new stdClass;
-            $ast->type = strtolower(str_ireplace(__NAMESPACE__.'\Element', '', get_class($this)));
+            $ast->type = strtolower(str_ireplace(Element::class.'\\', '', get_class($this)));
         }
 
         $this->ast = $ast;
@@ -64,7 +66,7 @@ abstract class Element implements JsonSerializable, ArrayAccess   {
 
         if (isset($ast->type)) {
 
-            $type = str_ireplace(__NAMESPACE__, '', $ast->type);
+            $type = $ast->type;
         }
 
         if ($type === '') {
@@ -72,7 +74,7 @@ abstract class Element implements JsonSerializable, ArrayAccess   {
             throw new InvalidArgumentException('Invalid ast provided');
         }
 
-        $className = __NAMESPACE__.'\Element'.$ast->type;
+        $className = Element::class.'\\'.ucfirst($ast->type);
         
 		return new $className($ast);
     }
@@ -165,31 +167,6 @@ abstract class Element implements JsonSerializable, ArrayAccess   {
         return $this->ast;
     }
 
-    public function offsetSet($offset, $value) {
-
-        if (is_callable([$this, 'set'.$offset])) {
-
-            call_user_func([$this, 'set'.$offset], $value);
-        }
-    }
-
-    public function offsetExists($offset) {
-        return is_callable([$this, 'get'.$offset]) || is_callable([$this, 'set'.$offset]);
-    }
-
-    public function offsetUnset($offset) {
-
-        if (is_callable([$this, 'set'.$offset])) {
-
-            call_user_func([$this, 'set'.$offset], null);
-        }
-    }
-
-    public function offsetGet($offset) {
-
-        return is_callable([$this, 'get'.$offset]) ? call_user_func([$this, 'get'.$offset]): null;
-    }
-
     /**
      * @return string
      * @throws Exception
@@ -198,13 +175,15 @@ abstract class Element implements JsonSerializable, ArrayAccess   {
     {
         try {
 
-            return (new Identity())->render($this, null, true);
+            return (new Renderer())->render($this, null, true);
         }
 
         catch (Exception $ex) {
 
-            echo $ex->getTraceAsString();
+            error_log($ex->getTraceAsString());
         }
+
+        return '';
     }
 
     public function __clone()
